feat(wechat-public-account-push): 新增好文节选
- 修改了常量的格式，更符合规范了！

- 应要求加入了文学类的出参
