+++ config/index.js
@@ -1,28 +1,51 @@
 export const config = {
-    // 公众号配置
-    // 公众号appId
-    appId: "",
-    // 公众号appSecret
-    appSecret: "",
+    /**
+     * 公众号配置
+     */
+
+    // 公众号APP_ID
+    APP_ID: "",
+    // 公众号APP_SECRET
+    APP_SECRET: "",
     // 模板消息id
-    templateId: "",
+    TEMPLATE_ID: "",
     // 接收公众号消息的微信号，如果有多个，需要在[]里用英文逗号间隔，例如["wx1", "wx2"]
-    user: [],
+    USERS: [""],
      
-    // 信息配置
+    /**
+     * 信息配置
+     */
+
+    /** 天气相关 */
+
     // 所在省份
-    province: "广东",
+    PROVINCE: "广东",
     // 所在城市
-    city: "惠州",
-    // 生日，修改名字为对应需要显示的名字，如果生日为农历， type为 r
-    birthdays: [
-      {"name": "老婆", "year": "1993", "date": "12-27", "type": 'new'},
-      {"name": "家公", "year": "1993", "date": "08-09", "type": 'new'},
+    CITY: "惠州",
+
+    /** 生日相关 */
+
+    // 生日，修改名字为对应需要显示的名字, data 仅填月日即可, 请严格按照示例填写
+    BIRTHDAYS: [
+      {"name": "老婆", "year": "1993", "date": "12-27"},
+      {"name": "家公", "year": "1993", "date": "08-09"},
     ],
+
+    /** 日期相关 */
+
     // 在一起的日子，格式同上
-    loveDate: "2015-05-01",
+    LOVE_DATE: "2015-05-01",
     // 结婚纪念日
-    marryDate: "2020-01-04"
+    MARRY_DATE: "2020-01-04",
+
+
+    /** 好文节选 */
+
+    // 好文节选的内容类型
+    // 可以填写【动画，漫画，游戏，小说，原创，网络，其他】； 随机则填写 ""
+    LITERARY_PREFERENCE: ""
+
+
     }
 
 // {{date.DATA}}
